requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==4.0.1
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
loguru==0.7.2
redis==5.0.1
psycopg2-binary==2.9.9
fake-useragent==1.4.0
tenacity==8.2.3
